Update #set('host', url) for missing ports.
Increases reliability of `.set('host', url)` for `url`s missing the port.
`hostname` propagation still happens when the port is not present.
Increases IPv6 support with brackets.

Skipped (failing) test case added - IPv6 without brackets will fail,
as it doesn't recognize if the address is shortened or not. A full
address with 8 octets will be parsed as 7.